requirements.txt
google-generativeai>=0.8.0
google-adk>=0.1.0
chromadb>=0.5.0
PyMuPDF>=1.24.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
sentence-transformers